requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
Django==4.1.3
django-celery-results==2.4.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.5
django-redis==5.2.0
django-smtp-ssl==1.0
djangorestframework==3.14.0
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
idna==3.4
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.32
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1